restructure scripts, configurable deps, add ABOUT file, update README
