+++ .gitignore
@@ -114,4 +114,4 @@ venv.bak/
 .bkp
 .zip
 geocatbridge/libs/*/
-dependencies/
+dependencies/*/
