+++ scripts/build.py
@@ -6,6 +6,7 @@ import argparse
 from pathlib import Path
 from fnmatch import fnmatch
 
+from shared import get_rootdir
 from geocatbridge.utils.meta import getVersion
 
 OUTPUT_DIRNAME = 'build'
@@ -17,7 +18,7 @@ def package():
     ns = ap.parse_args()
 
     file_name = f'{ns.filename}.zip'
-    output_dir = Path(f'./{OUTPUT_DIRNAME}').resolve()
+    output_dir = get_rootdir() / OUTPUT_DIRNAME
     os.makedirs(output_dir, exist_ok=True)
     package_file = output_dir / file_name
     with zipfile.ZipFile(package_file, "w", zipfile.ZIP_DEFLATED) as f:
@@ -28,9 +29,9 @@ def make_zip(zip_file):
     print("Creating ZIP for GeoCat Bridge plugin...")
     file_excludes = {'*.pyc', "*.git*", "*.log"}
     dir_excludes = {"test", "tests", "_debug", "__pycache__"}  # folder names inside ./geocatbridge
-    src_dir = f"./geocatbridge"
-    gpltxt = Path('./LICENSE')
-    readme = Path('./README.md')
+    root_dir = get_rootdir()
+    src_dir = root_dir / "geocatbridge"
+    gpltxt = root_dir / 'LICENSE'
 
     def filter_excludes(file_list):
         for fn in file_list:
@@ -43,13 +44,7 @@ def make_zip(zip_file):
         print(f'\t{dstpath}')
         zip_file.write(gpltxt.resolve(), dstpath)
     else:
-        raise FileNotFoundError(f"Required LICENSE file is missing!")
-
-    if readme.is_file():
-        # Include README file from repo root in package if it exists
-        dstpath = src_dir / readme
-        print(f'\t{dstpath}')
-        zip_file.write(readme.resolve(), dstpath)
+        raise FileNotFoundError(f"❌ Required LICENSE file is missing!")
 
     for root, dirs, files in os.walk(src_dir):
         if any(p in dir_excludes for p in Path(root).parts):
@@ -60,7 +55,7 @@ def make_zip(zip_file):
             print(f'\t{dstpath}')
             zip_file.write(os.path.join(root, f), dstpath)
 
-    print(f'Successfully created ZIP file {zip_file.filename}')
+    print(f'✅ Successfully created ZIP file {zip_file.filename}')
 
 
 if __name__ == "__main__":
